created separate admin panel item, styling change
