+++ src/index.css
@@ -191,3 +191,7 @@ body {
     filter: invert(1);
 }
 
+input[type="date"] {
+  color: rgb(124, 122, 122); /* Set your desired color */
+}
+
